# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Chocolate factory tours booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  /** Confirmation code */
  code: string;
  slotId: number;
  tourName: string;
  date: string;
  startTime: string;
  customerName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  peopleCount: number;
  pricePerPerson: number;
  totalPrice: number;
  status: BookingStatus;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}